Pong/requirements.txt
Django==5.0.4
psycopg==3.1.8
gunicorn==22.0.0
#python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
fontawesomefree==6.5.1
PyJWT==2.8.0
pyotp==2.9.0
pillow==10.3.0
requests==2.32.3
qrcode[pil]
channels
daphne
channels-redis